requirements.txt
babel==2.17.0
blinker==1.9.0
Bootstrap-Flask==2.5.0
cachelib==0.13.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
defusedxml==0.7.1
dnspython==2.8.0
Flask==3.1.2
flask-babel==4.0.0
Flask-Caching==2.3.1
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
ipwhois==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dateutil==2.9.0.post0
python-whois==0.9.6
pytz==2025.2
requests==2.32.5
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
